Добавлен рест для добавления строк, исправлена java heap spase, исправлен пересчет процентов, добавлено описание

+++ backend/src/main/java/ru/stm/delete_rows/service/strategy/RemoveStrategy.java
@@ -16,5 +16,5 @@ public interface RemoveStrategy {
      * @param percent процент удаляемых записей
      * @return true при соответсвии стратегии параметрам, иначе false
      */
-    boolean isRecommended(long percent);
+    boolean isRecommended(double percent);
 }
